しぬ

+++ app/crud.py
@@ -3,25 +3,22 @@
 from sqlalchemy.orm import Session
 from . import models, schemas
 
-# Todoの取得
 def get_todos(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.Todo).offset(skip).limit(limit).all()
 
-# Todoの削除
 def delete_todo(db: Session, todo_id: int):
     db_todo = db.query(models.Todo).filter(models.Todo.id == todo_id).first()
+    if not db_todo:
+        return None
     db.delete(db_todo)
     db.commit()
     return db_todo
 
-# crud.py に追加
-
 def get_todo_by_id(db: Session, todo_id: int):
     return db.query(models.Todo).filter(models.Todo.id == todo_id).first()
 
-
 def create_todo(db: Session, todo: schemas.TodoCreate):
-    db_todo = models.Todo(title=todo.title)  # completed はデフォルトで False
+    db_todo = models.Todo(title=todo.title)
     db.add(db_todo)
     db.commit()
     db.refresh(db_todo)
